+++ include/layer.h
@@ -22,15 +22,15 @@ typedef struct layer_t {
 
 layer_t* layer_init(int input_size, int output_size, activation_type act);
 void layer_forward(layer_t* layer, matrix_t* input);
-void layer_backwards(layer_t* layer, matrix_t* grad_output, float learning_rate);
+void layer_backwards(layer_t* layer, matrix_t* grad_output, double learning_rate);
 void layer_free(layer_t* layer);
 
 void init_bias(matrix_t* m, activation_type act);
 
 void init_weights(matrix_t* m, activation_type act, int inputs, int outputs);
-void init_weights_relu(matrix_t* m, int inputs, float mu, float sigma);
+void init_weights_relu(matrix_t* m, int inputs, double mu, double sigma);
 void init_weights_sigmoid(matrix_t* m, int inputs, int outputs);
 
 void init_randf_vals(matrix_t* m);
 
-float box_muller_tran(float mu, float sigma);
+double box_muller_tran(double mu, double sigma);
