+++ src/utils.c
@@ -1,10 +1,10 @@
 #include "../include/utils.h"
 #include <stdlib.h>
 
-float derivative(float (*p_act_func)(float), float x, float h_step){
+double derivative(double (*p_act_func)(double), double x, double h_step){
     return (p_act_func(x+h_step)-p_act_func(x-h_step)) / (2*h_step);
 }
 
-float randf(){
-    return (float)rand() / ((float)RAND_MAX+1.);
+double randf(){
+    return (double)rand() / ((double)RAND_MAX+1.);
 }
